Crate objc2_quartz_core

Bindings to the CoreAnimation framework

See Apple’s docs and the general docs on framework crates for more information.

This actually lives in the QuartzCore framework, but CoreAnimation is the name that people use to refer to it.

Structs

CAAutoresizingMask

Apple’s documentation

CACornerMask

Apple’s documentation

CADisplayLink

Class representing a timer bound to the display vsync. *

CAEdgeAntialiasingMask

Apple’s documentation

CAFrameRateRange

Apple’s documentation

CALayer

The base layer class. *

CATransform3D

Apple’s documentation

Statics

CAFrameRateRangeDefault⚠

Apple’s documentation

CAToneMapModeAutomatic⚠

Apple’s documentation

CAToneMapModeIfSupported⚠

Apple’s documentation

CAToneMapModeNever⚠

Apple’s documentation

CATransform3DIdentity⚠

Apple’s documentation

kCAContentsFormatGray8Uint⚠

Apple’s documentation

kCAContentsFormatRGBA8Uint⚠

Layer contentsFormat values. *

kCAContentsFormatRGBA16Float⚠

Apple’s documentation

kCACornerCurveCircular⚠

Corner curve names. *

kCACornerCurveContinuous⚠

Apple’s documentation

kCAFillModeBackwards⚠

Apple’s documentation

kCAFillModeBoth⚠

Apple’s documentation

kCAFillModeForwards⚠

Apple’s documentation

kCAFillModeRemoved⚠

Apple’s documentation

kCAFilterLinear⚠

Apple’s documentation

kCAFilterNearest⚠

Contents filter names. *

kCAFilterTrilinear⚠

Apple’s documentation

kCAGravityBottom⚠

Apple’s documentation

kCAGravityBottomLeft⚠

Apple’s documentation

kCAGravityBottomRight⚠

Apple’s documentation

kCAGravityCenter⚠

Layer `contentsGravity’ values. *

kCAGravityLeft⚠

Apple’s documentation

kCAGravityResize⚠

Apple’s documentation

kCAGravityResizeAspect⚠

Apple’s documentation

kCAGravityResizeAspectFill⚠

Apple’s documentation

kCAGravityRight⚠

Apple’s documentation

kCAGravityTop⚠

Apple’s documentation

kCAGravityTopLeft⚠

Apple’s documentation

kCAGravityTopRight⚠

Apple’s documentation

kCAOnOrderIn⚠

Layer event names. *

kCAOnOrderOut⚠

Apple’s documentation

kCATransition⚠

The animation key used for transitions. *

Traits

CAAction

Action (event handler) protocol. *

CALayerDelegate

Delegate methods. *

CALayoutManager

Layout manager protocol. *

CAMediaTiming

Apple’s documentation

NSValueCATransform3DAdditions

Category “CATransform3DAdditions” on NSValue.

Functions

CAFrameRateRangeIsEqualToRange⚠

CAFrameRateRangeMake⚠

CATransform3DConcat

CATransform3DEqualToTransform

CATransform3DGetAffineTransform

CATransform3DInvert

CATransform3DIsAffine

CATransform3DIsIdentity

CATransform3DMakeAffineTransform

CATransform3DMakeRotation

CATransform3DMakeScale

CATransform3DMakeTranslation

CATransform3DRotate

CATransform3DScale

CATransform3DTranslate

Type Aliases

CALayerContentsFilter

Apple’s documentation

CALayerContentsFormat

Apple’s documentation

CALayerContentsGravity

Apple’s documentation

CALayerCornerCurve

Apple’s documentation

CAMediaTimingFillMode

Apple’s documentation

CAToneMapMode

Apple’s documentation