Crate objc2_quartz_core

Bindings to the CoreAnimation framework

See Apple’s docs and the general docs on framework crates for more information.

This actually lives in the QuartzCore framework, but CoreAnimation is the name that people use to refer to it.

Structs

CAAnimation

The base animation class. *

CAAnimationGroup

Animation subclass for grouped animations. *

CAAutoresizingMask

Apple’s documentation

CABasicAnimation

Subclass for basic (single-keyframe) animations. *

CAConstraint

The class representing a single layout constraint. *

CAConstraintAttribute

Apple’s documentation

CAConstraintLayoutManager

The constraint-based layout manager. *

CACornerMask

Apple’s documentation

CADisplayLink

Class representing a timer bound to the display vsync. *

CAEDRMetadata

Apple’s documentation

CAEdgeAntialiasingMask

Apple’s documentation

CAEmitterCell

Apple’s documentation

CAEmitterLayer

Apple’s documentation

CAFrameRateRange

Apple’s documentation

CAGradientLayer

Apple’s documentation

CAKeyframeAnimation

General keyframe animation class. *

CALayer

The base layer class. *

CAMediaTimingFunction

Apple’s documentation

CAMetalDisplayLink

Apple’s documentation

CAMetalDisplayLinkUpdate

Apple’s documentation

CAMetalLayer

Apple’s documentation

CAOpenGLLayerDeprecated

Apple’s documentation

CAPropertyAnimation

Subclass for property-based animations. *

CARemoteLayerClient

Apple’s documentation

CARemoteLayerServer

Apple’s documentation

CARenderer

Apple’s documentation

CAReplicatorLayer

Apple’s documentation

CAScrollLayer

Apple’s documentation

CAShapeLayer

Apple’s documentation

CASpringAnimation

Subclass for mass-spring animations.

CATextLayer

Apple’s documentation

CATiledLayer

Apple’s documentation

CATransaction

Apple’s documentation

CATransform3D

Apple’s documentation

CATransformLayer

Apple’s documentation

CATransition

Transition animation subclass. *

CAValueFunction

Apple’s documentation

Statics

CADynamicRangeAutomatic⚠

Apple’s documentation

CADynamicRangeConstrainedHigh⚠

Apple’s documentation

CADynamicRangeHigh⚠

Apple’s documentation

CADynamicRangeStandard⚠

Apple’s documentation

CAFrameRateRangeDefault⚠

Apple’s documentation

CAToneMapModeAutomatic⚠

Apple’s documentation

CAToneMapModeIfSupported⚠

Apple’s documentation

CAToneMapModeNever⚠

Apple’s documentation

CATransform3DIdentity⚠

Apple’s documentation

kCAAlignmentCenter⚠

Apple’s documentation

kCAAlignmentJustified⚠

Apple’s documentation

kCAAlignmentLeft⚠

Apple’s documentation

kCAAlignmentNatural⚠

Apple’s documentation

kCAAlignmentRight⚠

Apple’s documentation

kCAAnimationCubic⚠

Apple’s documentation

kCAAnimationCubicPaced⚠

Apple’s documentation

kCAAnimationDiscrete⚠

Apple’s documentation

kCAAnimationLinear⚠

Apple’s documentation

kCAAnimationPaced⚠

Apple’s documentation

kCAAnimationRotateAuto⚠

Apple’s documentation

kCAAnimationRotateAutoReverse⚠

Apple’s documentation

kCAContentsFormatAutomatic⚠

Apple’s documentation

kCAContentsFormatGray8Uint⚠

Apple’s documentation

kCAContentsFormatRGBA8Uint⚠

Layer contentsFormat values. *

kCAContentsFormatRGBA16Float⚠

Apple’s documentation

kCACornerCurveCircular⚠

Corner curve names. *

kCACornerCurveContinuous⚠

Apple’s documentation

kCAEmitterLayerAdditive⚠

Apple’s documentation

kCAEmitterLayerBackToFront⚠

Apple’s documentation

kCAEmitterLayerCircle⚠

Apple’s documentation

kCAEmitterLayerCuboid⚠

Apple’s documentation

kCAEmitterLayerLine⚠

Apple’s documentation

kCAEmitterLayerOldestFirst⚠

Apple’s documentation

kCAEmitterLayerOldestLast⚠

Apple’s documentation

kCAEmitterLayerOutline⚠

Apple’s documentation

kCAEmitterLayerPoint⚠

`emitterShape’ values. *

kCAEmitterLayerPoints⚠

`emitterMode’ values. *

kCAEmitterLayerRectangle⚠

Apple’s documentation

kCAEmitterLayerSphere⚠

Apple’s documentation

kCAEmitterLayerSurface⚠

Apple’s documentation

kCAEmitterLayerUnordered⚠

`renderMode’ values. *

kCAEmitterLayerVolume⚠

Apple’s documentation

kCAFillModeBackwards⚠

Apple’s documentation

kCAFillModeBoth⚠

Apple’s documentation

kCAFillModeForwards⚠

Apple’s documentation

kCAFillModeRemoved⚠

Apple’s documentation

kCAFillRuleEvenOdd⚠

Apple’s documentation

kCAFillRuleNonZero⚠

Apple’s documentation

kCAFilterLinear⚠

Apple’s documentation

kCAFilterNearest⚠

Contents filter names. *

kCAFilterTrilinear⚠

Apple’s documentation

kCAGradientLayerAxial⚠

`type’ values. *

kCAGradientLayerConic⚠

Apple’s documentation

kCAGradientLayerRadial⚠

Apple’s documentation

kCAGravityBottom⚠

Apple’s documentation

kCAGravityBottomLeft⚠

Apple’s documentation

kCAGravityBottomRight⚠

Apple’s documentation

kCAGravityCenter⚠

Layer `contentsGravity’ values. *

kCAGravityLeft⚠

Apple’s documentation

kCAGravityResize⚠

Apple’s documentation

kCAGravityResizeAspect⚠

Apple’s documentation

kCAGravityResizeAspectFill⚠

Apple’s documentation

kCAGravityRight⚠

Apple’s documentation

kCAGravityTop⚠

Apple’s documentation

kCAGravityTopLeft⚠

Apple’s documentation

kCAGravityTopRight⚠

Apple’s documentation

kCALineCapButt⚠

Apple’s documentation

kCALineCapRound⚠

Apple’s documentation

kCALineCapSquare⚠

Apple’s documentation

kCALineJoinBevel⚠

Apple’s documentation

kCALineJoinMiter⚠

Apple’s documentation

kCALineJoinRound⚠

Apple’s documentation

kCAMediaTimingFunctionDefault⚠

Apple’s documentation

kCAMediaTimingFunctionEaseIn⚠

Apple’s documentation

kCAMediaTimingFunctionEaseInEaseOut⚠

Apple’s documentation

kCAMediaTimingFunctionEaseOut⚠

Apple’s documentation

kCAMediaTimingFunctionLinear⚠

Timing function names. *

kCAOnOrderIn⚠

Layer event names. *

kCAOnOrderOut⚠

Apple’s documentation

kCARendererColorSpace⚠

Options for the renderer options dictionary. *

kCARendererMetalCommandQueue⚠

Apple’s documentation

kCAScrollBoth⚠

Apple’s documentation

kCAScrollHorizontally⚠

Apple’s documentation

kCAScrollNone⚠

Apple’s documentation

kCAScrollVertically⚠

Apple’s documentation

kCATransactionAnimationDuration⚠

Transaction property ids. *

kCATransactionAnimationTimingFunction⚠

Apple’s documentation

kCATransactionCompletionBlock⚠

Apple’s documentation

kCATransactionDisableActions⚠

Apple’s documentation

kCATransition⚠

The animation key used for transitions. *

kCATransitionFade⚠

Apple’s documentation

kCATransitionFromBottom⚠

Apple’s documentation

kCATransitionFromLeft⚠

Apple’s documentation

kCATransitionFromRight⚠

Apple’s documentation

kCATransitionFromTop⚠

Apple’s documentation

kCATransitionMoveIn⚠

Apple’s documentation

kCATransitionPush⚠

Apple’s documentation

kCATransitionReveal⚠

Apple’s documentation

kCATruncationEnd⚠

Apple’s documentation

kCATruncationMiddle⚠

Apple’s documentation

kCATruncationNone⚠

Apple’s documentation

kCATruncationStart⚠

Apple’s documentation

kCAValueFunctionRotateX⚠

Value function names. *

kCAValueFunctionRotateY⚠

Apple’s documentation

kCAValueFunctionRotateZ⚠

Apple’s documentation

kCAValueFunctionScale⚠

Apple’s documentation

kCAValueFunctionScaleX⚠

Apple’s documentation

kCAValueFunctionScaleY⚠

Apple’s documentation

kCAValueFunctionScaleZ⚠

Apple’s documentation

kCAValueFunctionTranslate⚠

Apple’s documentation

kCAValueFunctionTranslateX⚠

Apple’s documentation

kCAValueFunctionTranslateY⚠

Apple’s documentation

kCAValueFunctionTranslateZ⚠

Apple’s documentation

Traits

CAAction

Action (event handler) protocol. *

CAAnimationDelegate

Apple’s documentation

CALayerDelegate

Delegate methods. *

CALayoutManager

Layout manager protocol. *

CAMediaTiming

Apple’s documentation

CAMetalDisplayLinkDelegate

Apple’s documentation

CAMetalDrawable

Apple’s documentation

NSValueCATransform3DAdditions

Category “CATransform3DAdditions” on NSValue.

Functions

CACurrentMediaTime

CAFrameRateRangeIsEqualToRangeDeprecated

CAFrameRateRangeMakeDeprecated

CATransform3DConcatDeprecated

CATransform3DEqualToTransformDeprecated

CATransform3DGetAffineTransformDeprecated

CATransform3DInvertDeprecated

CATransform3DIsAffineDeprecated

CATransform3DIsIdentityDeprecated

CATransform3DMakeAffineTransformDeprecated

CATransform3DMakeRotationDeprecated

CATransform3DMakeScaleDeprecated

CATransform3DMakeTranslationDeprecated

CATransform3DRotateDeprecated

CATransform3DScaleDeprecated

CATransform3DTranslateDeprecated

Type Aliases

CAAnimationCalculationMode

Apple’s documentation

CAAnimationRotationMode

Apple’s documentation

CADynamicRange

Apple’s documentation

CAEmitterLayerEmitterMode

Apple’s documentation

CAEmitterLayerEmitterShape

Apple’s documentation

CAEmitterLayerRenderMode

Apple’s documentation

CAGradientLayerType

Apple’s documentation

CALayerContentsFilter

Apple’s documentation

CALayerContentsFormat

Apple’s documentation

CALayerContentsGravity

Apple’s documentation

CALayerCornerCurve

Apple’s documentation

CAMediaTimingFillMode

Apple’s documentation

CAMediaTimingFunctionName

Apple’s documentation

CAScrollLayerScrollMode

Apple’s documentation

CAShapeLayerFillRule

Apple’s documentation

CAShapeLayerLineCap

Apple’s documentation

CAShapeLayerLineJoin

Apple’s documentation

CATextLayerAlignmentMode

Apple’s documentation

CATextLayerTruncationMode

Apple’s documentation

CAToneMapMode

Apple’s documentation

CATransitionSubtype

Apple’s documentation

CATransitionType

Apple’s documentation

CAValueFunctionName

Apple’s documentation